Move config.py to client dir
Having the library create a default empty `Config()` object doesn't actually do anything - it will always be empty, so we'll always need to pass in username / password via function parameters. And since the application always needs to pass data in via parameters, even if it _wasn't_ empty (eg if the user manually created config.json by hand), the library config would still be overridden by the client config.

Since the library config does nothing except add confusion, and the client config reads and writes and does all the work, I think it makes more sense for it to be part of the client - which also means that the library itself can be kept as a simpler pure network wrapper